Update Cauldron.gd
Here is my updated cauldron code. Incredibly messy and janky, will attempt to fix more soon, but for now it works. There is one big issue i noticed tho. If there is a potion on the cauldron, and you through something in the trash, the trashed item goes on the cauldron, and you pick up a yellow potion. Idk why that happens yet, but I will figure it out!